requirements.txt
django
requests
pytest
nltk==3.4
python-twitter == 3.5
